+++ GgafDx/include/jp/ggaf/dx/actor/supporter/VecVehicleMvAngAssistant.h
@@ -87,11 +87,11 @@ public:
      * @param prm_zero_acc_end_flg true:目標移動距離に達した際に角加速度を０に強制設定/false:角加速度はそのままにしておく
      */
     void turnRzByDt(angle prm_distance, int prm_target_frames,
-                    float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+                    double prm_p1, double prm_p2, angvelo prm_end_angvelo,
                     bool prm_zero_acc_end_flg);
 
     void turnRyByDt(angle prm_distance, int prm_target_frames,
-                    float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+                    double prm_p1, double prm_p2, angvelo prm_end_angvelo,
                     bool prm_zero_acc_end_flg);
     /**
      * 目標の移動方角(Z軸)へ滑らかに回転するシークエンスを実行(時間指定、角速度変動) .
@@ -105,7 +105,7 @@ public:
      * @param prm_zero_acc_end_flg true:目標移動距離に達した際に角加速度を０に強制設定/false:角加速度はそのままにしておく
      */
     void turnRzByDtTo(angle prm_rz_target, int prm_way, int prm_target_frames,
-                      float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+                      double prm_p1, double prm_p2, angvelo prm_end_angvelo,
                       bool prm_zero_acc_end_flg);
 
     /**
@@ -120,7 +120,7 @@ public:
      * @param prm_zero_acc_end_flg true:目標移動距離に達した際に角加速度を０に強制設定/false:角加速度はそのままにしておく
      */
     void turnRyByDtTo(angle prm_ang_ry_target, int prm_way, int prm_target_frames,
-                      float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+                      double prm_p1, double prm_p2, angvelo prm_end_angvelo,
                       bool prm_zero_acc_end_flg);
 
     /**
@@ -146,7 +146,7 @@ public:
     void turnRzRyByDtTo(
             angle prm_rz_target, angle prm_ry_target, int prm_way, bool prm_optimize_ang,
             int prm_target_frames,
-            float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+            double prm_p1, double prm_p2, angvelo prm_end_angvelo,
             bool prm_zero_acc_end_flg);
 
     /**
@@ -166,7 +166,7 @@ public:
     void turnByDtTwd(
             coord prm_tx, coord prm_ty, coord prm_tz, int prm_way, bool prm_optimize_ang,
             int prm_target_frames,
-            float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+            double prm_p1, double prm_p2, angvelo prm_end_angvelo,
             bool prm_zero_acc_end_flg);
 
     /**
@@ -184,7 +184,7 @@ public:
     void turnByDtTwd(
             GeometricActor* prm_pActor_target, int prm_way, bool prm_optimize_ang,
             int prm_target_frames,
-            float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+            double prm_p1, double prm_p2, angvelo prm_end_angvelo,
             bool prm_zero_acc_end_flg);
 
     /**
@@ -225,40 +225,40 @@ public:
      */
     void turnRzByVd(
             angvelo prm_top_angvelo, angle prm_distance,
-            float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+            double prm_p1, double prm_p2, angvelo prm_end_angvelo,
             bool prm_zero_acc_end_flg);
 
     void turnRyByVd(
             angvelo prm_top_angvelo, angle prm_distance,
-            float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+            double prm_p1, double prm_p2, angvelo prm_end_angvelo,
             bool prm_zero_acc_end_flg);
 
     void turnRzByVdTo(
             angvelo prm_top_angvelo, angle prm_rz_target, int prm_way,
-            float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+            double prm_p1, double prm_p2, angvelo prm_end_angvelo,
             bool prm_zero_acc_end_flg);
 
     void turnRyByVdTo(
             angvelo prm_top_angvelo, angle prm_ry_target, int prm_way,
-            float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+            double prm_p1, double prm_p2, angvelo prm_end_angvelo,
             bool prm_zero_acc_end_flg);
 
     void turnRzRyByVdTo(
             angvelo prm_top_angvelo,
             angle prm_rz_target, angle prm_ry_target, int prm_way, bool prm_optimize_ang,
-            float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+            double prm_p1, double prm_p2, angvelo prm_end_angvelo,
             bool prm_zero_acc_end_flg);
 
     void turnByVdTwd(
             angvelo prm_top_angvelo,
             coord prm_tx, coord prm_ty, coord prm_tz, int prm_way, bool prm_optimize_ang,
-            float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+            double prm_p1, double prm_p2, angvelo prm_end_angvelo,
             bool prm_zero_acc_end_flg);
 
     void turnByVdTwd(
             angvelo prm_top_angvelo,
             GeometricActor* prm_pActor_target, int prm_way, bool prm_optimize_ang,
-            float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+            double prm_p1, double prm_p2, angvelo prm_end_angvelo,
             bool prm_zero_acc_end_flg);
 
 
@@ -266,13 +266,13 @@ public:
     void turnRzPendulum(coord prm_target_rz1, coord prm_target_rz2,
                         int prm_twist_num,
                         int prm_first_way, int prm_target_frames,
-                        float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+                        double prm_p1, double prm_p2, angvelo prm_end_angvelo,
                         bool prm_zero_acc_end_flg);
 
     void turnRyPendulum(coord prm_target_ry1, coord prm_target_ry2,
                         int prm_twist_num,
                         int prm_first_way, int prm_target_frames,
-                        float prm_p1, float prm_p2, angvelo prm_end_angvelo,
+                        double prm_p1, double prm_p2, angvelo prm_end_angvelo,
                         bool prm_zero_acc_end_flg);
